[level 3] Title: 디스크 컨트롤러, Time: 6.58 ms, Memory: 90.5 MB -BaekjoonHub

+++ 프로그래머스/3/42627. 디스크 컨트롤러/디스크 컨트롤러.java
@@ -33,6 +33,6 @@ class Solution {
             }
         }
         
-        return (int) sumOfTurnaroundTime / jobs.length;
+        return sumOfTurnaroundTime / jobs.length;
     }
 }
